+++ webclient/src/components/MainLayout/index.js
@@ -1,12 +1,25 @@
-import React, { Component } from "react";
+import React from "react";
 
-//import Product from "./../product";
+import { BrowserRouter as Router, Route, Link } from "react-router-dom";
+
+import Home from "./../Home";
+import Product from "./../product";
 import Purchase from "./../purchase";
 
-class MainLayout extends Component {
-  render() {
-    return <Purchase />;
-  }
-}
+export default function MainLayout() {
+  return (
+    <Router>
+      <div>
+        <Link to="/">Home</Link> &nbsp;
+        <Link to="/product">Product</Link>  &nbsp;
+        <Link to="/purchase">Purchase</Link>
 
-export default MainLayout;
+        <hr />
+
+        <Route exact path="/" component={Home} />
+        <Route path="/product" component={Product} />
+        <Route path="/purchase" component={Purchase} />
+      </div>
+    </Router>
+  );
+}