+++ webclient/src/components/product/index.js
@@ -1,12 +1,5 @@
-import React, { useState } from "react";
-
-//Libs
-
-//style component
-import { Row, Column } from "./../style/grid";
-import { Button } from "./../style/form";
-
-//local
+import React from "react";
+import { Route, Link } from "react-router-dom";
 
 //Private
 import ProductView from "./ProductView";
@@ -14,54 +7,25 @@ import ProductList from "./ProductList";
 import ProductFormEdit from "./ProductFormEdit";
 import ProductFormNew from "./ProductFormNew";
 
-export default function Product() {
-  const [selectedProductId, setSelectedProductId] = useState(0);
-  const [mode, setMode] = useState("");
-
-  const ChangeModeToView = id => {
-    setMode("view");
-    setSelectedProductId(id);
-  };
-
-  const ChangeModeToEdit = id => {
-    setMode("edit");
-    setSelectedProductId(id);
-  };
-
-  const ChangeModeToAdd = () => {
-    setMode("new");
-  };
-
-  let View = null;
-
-  if (mode === "view") {
-    View = <ProductView ProductId={selectedProductId} />;
-  } else if (mode === "new") {
-    View = <ProductFormNew />;
-  } else if (mode === "edit") {
-    View = <ProductFormEdit ProductId={selectedProductId} />;
-  } else {
-    View = <div>"select" </div>;
-  }
-
+export default function Product({ match }) {
   return (
-    <React.Fragment>
-      <Row>
-        <Column span="4">
-          <Button primary onClick={ChangeModeToAdd}>
-            Add new
-          </Button>
-        </Column>
-      </Row>
-      <Row>
-        <Column span="4">{View}</Column>
-        <Column span="8">
-          <ProductList
-            ViewClick={ChangeModeToView}
-            EditClick={ChangeModeToEdit}
-          />
-        </Column>
-      </Row>
-    </React.Fragment>
+    <div>
+      <b>Product</b> &nbsp;
+      <Link to={`${match.url}/list`}>List all</Link> &nbsp;
+      <Link to={`${match.url}/new`}>Add new</Link>
+      <hr />
+      <Route path={`${match.url}/list`} component={ProductList} />
+      <Route path={`${match.url}/new`} component={ProductFormNew} />
+      <Route path={`${match.url}/view/:ProductId`} component={ProductView} />
+      <Route
+        path={`${match.url}/edit/:ProductId`}
+        component={ProductFormEdit}
+      />
+      <Route
+        exact
+        path={match.url}
+        render={() => <h3>Please select an option</h3>}
+      />
+    </div>
   );
 }